requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==20.3.0             # via pytest
certifi==2020.11.8        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via uvicorn
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via networkx
fastapi==0.61.2           # via -r requirements.in
filelock==3.0.12          # via gdown
future==0.18.2            # via torch
gdown==3.11.1             # via -r requirements.in
gluoncv==0.7.0            # via -r requirements.in
graphviz==0.8.4           # via mxnet
h11==0.11.0               # via uvicorn
idna==2.10                # via requests
imageio==2.9.0            # via scikit-image
iniconfig==1.1.1          # via pytest
kiwisolver==1.3.1         # via matplotlib
matplotlib==3.3.3         # via gluoncv, scikit-image
mxnet==1.6.0              # via -r requirements.in
networkx==2.5             # via scikit-image
numpy==1.19.4             # via gluoncv, imageio, matplotlib, mxnet, opencv-contrib-python, pywavelets, scipy, torch, torchvision
opencv_contrib_python==4.2.0.34  # via -r requirements.in
packaging==20.4           # via pytest
pillow==7.1.2             # via -r requirements.in, gluoncv, imageio, matplotlib, scikit-image, torchvision
pluggy==0.13.1            # via pytest
portalocker==2.0.0        # via gluoncv
py==1.9.0                 # via pytest
pydantic==1.7.2           # via fastapi
pyparsing==2.4.7          # via matplotlib, packaging
pysocks==1.7.1            # via requests
pytest==6.1.2             # via -r requirements.in
python-dateutil==2.8.1    # via matplotlib
python-multipart==0.0.5   # via -r requirements.in
pywavelets==1.1.1         # via scikit-image
requests[socks]==2.25.0   # via gdown, gluoncv, mxnet
scikit_image==0.15.0      # via -r requirements.in
scipy==1.4.1              # via -r requirements.in, gluoncv, scikit-image
six==1.15.0               # via cycler, gdown, packaging, python-dateutil, python-multipart, torchvision
starlette==0.13.6         # via fastapi
toml==0.10.2              # via pytest
torch==1.5.0              # via -r requirements.in, torchvision
torchvision==0.2.2.post3  # via -r requirements.in
tqdm==4.43.0              # via -r requirements.in, gdown, gluoncv
urllib3==1.26.2           # via requests
uvicorn==0.12.3           # via -r requirements.in
